# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: c0ntr0l API
  version: 0.1.0
paths:
...
  /v1/chat/completions:
    post:
      tags:
      - chat
      summary: Create Chat Completion
      operationId: createChatCompletion
      security:
      - bearerAuth: []
      parameters:
      - name: X-Request-Id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 128
          - type: 'null'
          title: X-Request-Id
      - name: X-Control-Demo-Scenario
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 32
          - type: 'null'
          title: X-Control-Demo-Scenario
      - name: X-Control-Application
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 63
          - type: 'null'
          title: X-Control-Application
      - name: X-Control-Agent
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 63
          - type: 'null'
          title: X-Control-Agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletion'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/events:
    get:
      tags:
      - events
      summary: Stream Control Events
      operationId: streamControlEvents
      security:
      - bearerAuth: []
      parameters:
      - name: Last-Event-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^\d+-\d+$
          - type: 'null'
          title: Last-Event-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/runtime/executions:
    post:
      tags:
      - runtime
      summary: Start Runtime Execution
      operationId: startRuntimeExecution
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeExecutionRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeExecutionCreated'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /api/v1/runtime/executions/{execution_id}/preflight:
    post:
      tags:
      - runtime
      summary: Preflight Runtime Model Call
      operationId: preflightRuntimeModelCall
      security:
      - bearerAuth: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimePreflightRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimePreflightResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runtime/executions/{execution_id}/actions/check:
    post:
      tags:
      - runtime
      summary: Check Runtime Action
      operationId: checkRuntimeAction
      security:
      - bearerAuth: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeActionCheckRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeActionDecision'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runtime/executions/{execution_id}/actions/{action_id}/complete:
    post:
      tags:
      - runtime
      summary: Complete Runtime Action
      operationId: completeRuntimeAction
      security:
      - bearerAuth: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      - name: action_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Action Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeActionCompleteRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeActionCompleted'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runtime/executions/{execution_id}/intervention:
    get:
      tags:
      - runtime
      summary: Get Runtime Intervention
      operationId: getRuntimeIntervention
      security:
      - bearerAuth: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeIntervention'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runtime/executions/{execution_id}/cancel:
    post:
      tags:
      - runtime
      summary: Cancel Runtime Execution
      operationId: cancelRuntimeExecution
      security:
      - bearerAuth: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeCancellationResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runtime/executions/{execution_id}/recover:
    post:
      tags:
      - runtime
      summary: Recover Runtime Execution
      operationId: recoverRuntimeExecution
      security:
      - bearerAuth: []
      parameters:
      - name: execution_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Execution Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeRecoveryRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeRecoveryResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    ChatChoice:
      properties:
        index:
          type: integer
          title: Index
          default: 0
        message:
          $ref: '#/components/schemas/ResponseMessage'
        finish_reason:
          type: string
          title: Finish Reason
      type: object
      required:
      - message
      - finish_reason
      title: ChatChoice
    ChatCompletion:
      properties:
        id:
          type: string
          title: Id
        object:
          type: string
          title: Object
          default: chat.completion
        created:
          type: integer
          title: Created
        model:
          type: string
          title: Model
        choices:
          items:
            $ref: '#/components/schemas/ChatChoice'
          type: array
          minItems: 1
          title: Choices
        usage:
          $ref: '#/components/schemas/Usage'
      type: object
      required:
      - id
      - created
      - model
      - choices
      - usage
      title: ChatCompletion
    ChatMessage:
      properties:
        role:
          $ref: '#/components/schemas/MessageRole'
        content:
          type: string
          maxLength: 32768
          minLength: 1
          title: Content
        name:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Name
      type: object
      required:
      - role
      - content
      title: ChatMessage
    ChatRequest:
      properties:
        model:
          type: string
          maxLength: 128
          minLength: 1
          title: Model
        messages:
          items:
            $ref: '#/components/schemas/ChatMessage'
          type: array
          maxItems: 100
          minItems: 1
          title: Messages
        stream:
          type: boolean
          title: Stream
          default: false
        stream_options:
          anyOf:
          - $ref: '#/components/schemas/StreamOptions'
          - type: 'null'
        max_tokens:
          type: integer
          maximum: 4096.0
          minimum: 1.0
          title: Max Tokens
          default: 256
        temperature:
          anyOf:
          - type: number
            maximum: 2.0
            minimum: 0.0
          - type: 'null'
          title: Temperature
      type: object
      required:
      - model
      - messages
      title: ChatRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RecoveryStrategy:
      type: string
      enum:
      - retry_modified
      - model_handoff
      - manual_resume
      - stop
      title: RecoveryStrategy
...
    RuntimeActionCheckRequest:
      properties:
        kind:
          type: string
          pattern: ^(tool|model)$
          title: Kind
          default: tool
        name:
          type: string
          maxLength: 128
          minLength: 1
          title: Name
        arguments:
          additionalProperties: true
          type: object
          title: Arguments
      additionalProperties: false
      type: object
      required:
      - name
      title: RuntimeActionCheckRequest
    RuntimeActionCompleteRequest:
      properties:
        status:
          type: string
          pattern: ^(completed|failed)$
          title: Status
          default: completed
        result:
          title: Result
        progress:
          type: boolean
          title: Progress
        summary:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Summary
      additionalProperties: false
      type: object
      required:
      - progress
      title: RuntimeActionCompleteRequest
    RuntimeActionCompleted:
      properties:
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        action_id:
          type: string
          format: uuid
          title: Action Id
        status:
          type: string
          title: Status
        result_fingerprint:
          type: string
          title: Result Fingerprint
        progress:
          type: boolean
          title: Progress
      type: object
      required:
      - execution_id
      - action_id
      - status
      - result_fingerprint
      - progress
      title: RuntimeActionCompleted
    RuntimeActionDecision:
      properties:
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        action_id:
          type: string
          format: uuid
          title: Action Id
        decision:
          $ref: '#/components/schemas/RuntimeDecision'
        operation_fingerprint:
          type: string
          title: Operation Fingerprint
        occurrence:
          type: integer
          title: Occurrence
        threshold:
          type: integer
          title: Threshold
        reason:
          type: string
          title: Reason
        evidence:
          additionalProperties: true
          type: object
          title: Evidence
        checkpoint_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Checkpoint Id
      type: object
      required:
      - execution_id
      - action_id
      - decision
      - operation_fingerprint
      - occurrence
      - threshold
      - reason
      title: RuntimeActionDecision
...
    RuntimeCancellationResult:
      properties:
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        status:
          type: string
          title: Status
        checkpoint_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Checkpoint Id
      type: object
      required:
      - execution_id
      - status
      title: RuntimeCancellationResult
    RuntimeCheckpoint:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        status:
          type: string
          title: Status
        content_fingerprint:
          type: string
          title: Content Fingerprint
        packet:
          $ref: '#/components/schemas/ContinuityPacket'
        created_at:
          type: string
          format: date-time
          title: Created At
        consumed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Consumed At
      type: object
      required:
      - id
      - execution_id
      - status
      - content_fingerprint
      - packet
      - created_at
      title: RuntimeCheckpoint
    RuntimeDecision:
      type: string
      enum:
      - allow
      - warn
      - block
      - cancel
      title: RuntimeDecision
    RuntimeExecutionCreated:
      properties:
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        trace_id:
          type: string
          title: Trace Id
        status:
          type: string
          title: Status
        repeat_threshold:
          type: integer
          title: Repeat Threshold
        policy_mode:
          $ref: '#/components/schemas/RuntimePolicyMode'
      type: object
      required:
      - execution_id
      - trace_id
      - status
      - repeat_threshold
      - policy_mode
      title: RuntimeExecutionCreated
    RuntimeExecutionRequest:
      properties:
        task:
          type: string
          maxLength: 2000
          minLength: 1
          title: Task
        model:
          type: string
          maxLength: 128
          minLength: 1
          title: Model
        provider:
          type: string
          maxLength: 64
          minLength: 1
          title: Provider
          default: custom
        application_slug:
          anyOf:
          - type: string
            maxLength: 63
            minLength: 1
          - type: 'null'
          title: Application Slug
        agent_slug:
          anyOf:
          - type: string
            maxLength: 63
            minLength: 1
          - type: 'null'
          title: Agent Slug
        repeat_threshold:
          type: integer
          maximum: 20.0
          minimum: 2.0
          title: Repeat Threshold
          default: 3
        window_size:
          type: integer
          maximum: 100.0
          minimum: 4.0
          title: Window Size
          default: 12
        policy_mode:
          $ref: '#/components/schemas/RuntimePolicyMode'
          default: enforce
      additionalProperties: false
      type: object
      required:
      - task
      - model
      title: RuntimeExecutionRequest
    RuntimeIntervention:
      properties:
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        execution_status:
          type: string
          title: Execution Status
        policy_code:
          type: string
          title: Policy Code
        policy_mode:
          $ref: '#/components/schemas/RuntimePolicyMode'
        outcome:
          type: string
          title: Outcome
        reason:
          type: string
          title: Reason
        evidence:
          additionalProperties: true
          type: object
          title: Evidence
        decided_at:
          type: string
          format: date-time
          title: Decided At
        checkpoint:
          anyOf:
          - $ref: '#/components/schemas/RuntimeCheckpoint'
          - type: 'null'
      type: object
      required:
      - execution_id
      - execution_status
      - policy_code
      - policy_mode
      - outcome
      - reason
      - decided_at
      title: RuntimeIntervention
    RuntimePolicyMode:
      type: string
      enum:
      - observe
      - warn
      - enforce
      title: RuntimePolicyMode
    RuntimePreflightRequest:
      properties:
        input_tokens:
          type: integer
          minimum: 0.0
          title: Input Tokens
        requested_output_tokens:
          type: integer
          maximum: 32768.0
          minimum: 1.0
          title: Requested Output Tokens
        estimated_cost:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Estimated Cost
          default: '0'
      additionalProperties: false
      type: object
      required:
      - input_tokens
      - requested_output_tokens
      title: RuntimePreflightRequest
    RuntimePreflightResult:
      properties:
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        decision:
          $ref: '#/components/schemas/RuntimeDecision'
        reason:
          type: string
          title: Reason
        provider:
          type: string
          title: Provider
        model:
          type: string
          title: Model
        input_tokens:
          type: integer
          title: Input Tokens
        reserved_output_tokens:
          type: integer
          title: Reserved Output Tokens
        safety_margin_tokens:
          type: integer
          title: Safety Margin Tokens
        projected_context_tokens:
          type: integer
          title: Projected Context Tokens
        context_window_tokens:
          type: integer
          title: Context Window Tokens
        context_remaining_tokens:
          type: integer
          title: Context Remaining Tokens
        context_utilization:
          type: number
          minimum: 0.0
          title: Context Utilization
        budgets:
          items:
            $ref: '#/components/schemas/RuntimeBudgetProjection'
          type: array
          title: Budgets
        checkpoint_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Checkpoint Id
      type: object
      required:
      - execution_id
      - decision
      - reason
      - provider
      - model
      - input_tokens
      - reserved_output_tokens
      - safety_margin_tokens
      - projected_context_tokens
      - context_window_tokens
      - context_remaining_tokens
      - context_utilization
      title: RuntimePreflightResult
    RuntimeRecoveryRequest:
      properties:
        strategy:
          $ref: '#/components/schemas/RecoveryStrategy'
        target_provider:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Target Provider
        target_model:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Target Model
        modified_arguments:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Modified Arguments
      additionalProperties: false
      type: object
      required:
      - strategy
      title: RuntimeRecoveryRequest
    RuntimeRecoveryResult:
      properties:
        source_execution_id:
          type: string
          format: uuid
          title: Source Execution Id
        strategy:
          $ref: '#/components/schemas/RecoveryStrategy'
        status:
          type: string
          title: Status
        resumed_execution_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Resumed Execution Id
        target_provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Provider
        target_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Model
        checkpoint:
          $ref: '#/components/schemas/RuntimeCheckpoint'
        message:
          type: string
          title: Message
        completion:
          anyOf:
          - $ref: '#/components/schemas/ChatCompletion'
          - type: 'null'
      type: object
      required:
      - source_execution_id
      - strategy
      - status
      - checkpoint
      - message
      title: RuntimeRecoveryResult
...
    StreamOptions:
      properties:
        include_usage:
          type: boolean
          title: Include Usage
          default: true
      type: object
      title: StreamOptions
...
    Usage:
      properties:
        prompt_tokens:
          type: integer
          minimum: 0.0
          title: Prompt Tokens
        completion_tokens:
          type: integer
          minimum: 0.0
          title: Completion Tokens
        total_tokens:
          type: integer
          minimum: 0.0
          title: Total Tokens
      type: object
      required:
      - prompt_tokens
      - completion_tokens
      - total_tokens
      title: Usage
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer